Updated Main Scene
Added Empty GameObject to Hierarchy and renamed it Graph.
Added the Graph script component to the Graph game object.
Created a Cube game object in the Hierarchy and renamed it Point.
Applied the Point shader to the Point material.
Applied the Point materials to the Point game object.
Created a prefab of the Point game object.
Adjusted Inspector values on the Graph script.